requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
streamlit==1.28.1
pymongo==4.6.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==41.0.7
opencv-python==4.8.1.78
deepface==0.0.79
mediapipe==0.10.8
numpy==1.24.3
pandas==2.0.3
Pillow==10.1.0
requests==2.31.0
tensorflow==2.13.0
python-dotenv==1.0.0